final/faust_app/requirements.txt
# ставим faust-streaming без зависимостей, чтобы он не притащил mode-streaming 0.4.x
faust-streaming==0.10.13

# совместимые версии (важно!)
mode==4.3.2
mode-streaming==0.3.1

# стек клиента Kafka под faust-streaming
aiokafka==0.8.1
kafka-python==2.0.2

# совместимые сетевые либы
yarl==1.9.2
aiohttp==3.9.5
click==8.1.7

typing-extensions==4.9.0
uvloop==0.19.0 ; platform_system!="Windows"